Rename and refactor bug regression classes and files for clarity and consistency
